+++ zerver/lib/events.py
@@ -647,8 +647,8 @@ def fetch_initial_state_data(
     if want("stop_words"):
         state["stop_words"] = read_stop_words()
 
-    if want("update_display_settings") and not user_settings_object:
-        for prop in UserProfile.display_settings_legacy:
+    if want("update_preferences") and not user_settings_object:
+        for prop in UserProfile.preferences_legacy:
             state[prop] = getattr(settings_user, prop)
         state["emojiset_choices"] = UserProfile.emojiset_choices()
         state["timezone"] = canonicalize_timezone(settings_user.timezone)
@@ -1412,9 +1412,9 @@ def apply_event(
             state["realm_linkifiers"] = event["realm_linkifiers"]
     elif event["type"] == "realm_playgrounds":
         state["realm_playgrounds"] = event["realm_playgrounds"]
-    elif event["type"] == "update_display_settings":
+    elif event["type"] == "update_preferences":
         if event["setting_name"] != "timezone":
-            assert event["setting_name"] in UserProfile.display_settings_legacy
+            assert event["setting_name"] in UserProfile.preferences_legacy
         state[event["setting_name"]] = event["setting"]
     elif event["type"] == "update_global_notifications":
         assert event["notification_name"] in UserProfile.notification_settings_legacy
@@ -1425,7 +1425,7 @@ def apply_event(
         if event["property"] != "timezone":
             assert event["property"] in UserProfile.property_types
         if event["property"] in {
-            **UserProfile.display_settings_legacy,
+            **UserProfile.preferences_legacy,
             **UserProfile.notification_settings_legacy,
         }:
             state[event["property"]] = event["value"]